Adhere to DTU's license requirements
Apparently the model file that is used by the original TMHMM cannot be distributed with tmhmm.py due to DTU's license. DTU has requested the model to be removed. This will destroy a lot of existing workflows, annoy a lot of people, and turn a lot of science non-reproducible, but here we go.